refactor: extract vaults lib to clientSupport
